// src/enums/RoleName.ts
export enum RoleName {
  Kẻ_mạo_danh = 'Kẻ_mạo_danh',
  Ma_sói = 'Ma_sói',
  Kẻ_phản_bội = 'Kẻ_phản_bội',
  Thợ_hồ = 'Thợ_hồ',
  Tiên_tri = 'Tiên_tri',
  Đạo_tặc = 'Đạo_tặc',
  Kẻ_phá_hoại = 'Kẻ_phá_hoại',
  Bợm_nhậu = 'Bợm_nhậu',
  Cú_đêm = 'Cú_đêm',
  Dân_làng = 'Dân_làng',
  Thợ_săn = 'Thợ_săn',
  Kẻ_chán_đời = 'Kẻ_chán_đời'
}